@@ -1,16 +1,15 @@
 ﻿using AutoFixture;
 
-namespace EvilBaschdi.Testing
+namespace EvilBaschdi.Testing;
+
+/// <inheritdoc />
+internal class OmitAutoPropertiesTrueCompositeCustomization : CompositeCustomization
 {
-    /// <inheritdoc />
-    internal class OmitAutoPropertiesTrueCompositeCustomization : CompositeCustomization
+    /// <summary>
+    ///     Constructor
+    /// </summary>
+    public OmitAutoPropertiesTrueCompositeCustomization(ICustomization customization)
+        : base(new OmitAutoPropertiesTrueCustomization(), customization)
     {
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        public OmitAutoPropertiesTrueCompositeCustomization(ICustomization customization)
-            : base(new OmitAutoPropertiesTrueCustomization(), customization)
-        {
-        }
     }
 }
