@@ -1,10 +1,9 @@
 ﻿using AutoFixture;
 
-namespace EvilBaschdi.Testing
+namespace EvilBaschdi.Testing;
+
+/// <inheritdoc />
+internal class OmitAutoPropertiesTrueCustomization : ICustomization
 {
-    /// <inheritdoc />
-    internal class OmitAutoPropertiesTrueCustomization : ICustomization
-    {
-        public void Customize(IFixture fixture) => fixture.OmitAutoProperties = true;
-    }
+    public void Customize(IFixture fixture) => fixture.OmitAutoProperties = true;
 }
